+++ esvg/Polyline.cpp
@@ -61,7 +61,7 @@ esvg::render::Path esvg::Polyline::createPath() {
 	esvg::render::Path out;
 	out.clear();
 	out.moveTo(false, m_listPoint[0]);
-	for( int32_t iii=1; iii< m_listPoint.size(); iii++) {
+	for(size_t iii=1; iii< m_listPoint.size(); iii++) {
 		out.lineTo(false, m_listPoint[iii]);
 	}
 	out.stop();
